Added generic raster data type: GenRaster[T]
This is based on the specialized Numeric class. Currently doesn't have
any operations defined on it but it does have some benchmarks. The plan
is to have a parallel set of generic operations.

There's still a lot to do.

Also need to get the compiler plugin working properly... right now
it works but has debugging code which I'm having a hard time disabling.