Update ARG binary format to screen row order.
The old ARG binary format stored rows starting from (map coordinate) ymin
and increased to ymax, which is the opposite order of rows as displayed on
the screen. The new format starts at ymax and finishes at ymin, which reflects
the order of the rows as displayed on the screen.

See https://github.com/azavea/geotrellis/wiki/ARG-Specification for more information.