Benchmarks updated and broken into separate benchmarks.
The benchmarks are updated to use LoadRaster and the demo catalog which is
set to statically cache.  The Benchmarks are broken up into smaller runnable
benchmarks, and the longer running benchmarks are removed, as we need to not
comment in and out benchmarks (as the caliper benchmarking site is not
compatable with that approach).