Improve GeoTiff support
This commit introduces our own GeoTiff encoder. It only supports
the parts of GeoTiff that we are currently using, although it should
be easily extensible. In particular, it writes signed or unsigned
rasters with 8-, 16- or 32-bit values.

It should be possible to write out own GeoTiff decoder as well,
which would remove out geotools dependency.