Refactored PNG encoder.
This commit moves the PNG code to a model similar to the one we introduced
with GeoTIFF: a package containing an Encoder, Settings, and other data
types which conform closely to the PNG spec.

In addition, it adds the capability of rendering PNG files without using a
filter (previously, the Paeth filter was hardcoded).