Optimize unary local raster operations.
This checkin removes the old UnaryLocal machinery that was causing significant
slowdown through the introduction of int boxing.  Instead, the unary local operations
(which are quite simple) simply extend OpN.

The Identity operation has been removed, as it is redundant given CopyRaster.

A compiler bug forced the geotrellis map algebra style implicits to be moved to
geotrellis.Implicits.