Move all methods assuming indexable data to ArrayRasterData.
All methods that assume that the underlying raster data can be indexed
via an integer (and that there are less than 2.1 billion cells) have been
moved from RasterData to a new ArrayRasterData class.  RasterData's array
like methods (apply, update) have been moved to ArrayRasterData.  RasterData
has a new method asArray that returns an ArrayRasterData from a given RasterData,
but use of that method in the future should be hidden behind code that handles
rasters that can not or should not be loaded into an in-memory array.