Introduce basic TiledRasterData support, and related operations.
This commit introduces TiledRasterData, our first RasterData class that doesn't
extend ArrayRasterData. This means that actions on it need to be defined in
terms of traversing operations like map/foreach, or other kinds of
transformations, rather than treating it like an array.

In addition to changes to Raster/RasterData, the commit also introduces some
operations designed for tiled raster data.