Fix PNG rendering in the demo.
Unfortunately, this commit encompasses a bunch of work. While moving the demo
code to a subproject (involving a lot of changes) I noticed it's PNG rendering
was broken, and stashed the work. This commit encompasses that change as well
as the work needed to fix PNG rendering again.

This includes a tweak to the LazyConverter raster data which prevents things
like map from exploding.