Rename geotrellis.operation package to geotrellis.op.
The geotrellis.operation package is renamed as geotrellis.op in this
checkin, which touches a lot of files.

A related change in this checkin is that the package name of all operations
now matches the directory name of that operation, which necessitates many
new import statements throughout the code base.