Reorganize packages for clearer imports.
This is a major refactor in which the geotrellis.op package is broken up
into many smaller op packages in the subdirectories related to the operations
at hand.  This has a number of very nice implications for developers using the
operations packages, and addresses some issues that had arisen from our previous
reorganization (especially regarding package namespace collision).

A few core packages are moved up into the geotrellis package, so that a single
"import geotrellis._" is sufficient for many simple operations.