Flesh out local map algebra operations.
This commit renames some existing operations, and adds a bunch more. The goal
is to support the "standard suite" of programming operators, i.e.:

 + - * / ~ & | ^ < <= > >= == != min max

This commit does not add symbolic operators (e.g. ==), but just adds the actual
named operations (e.g. Equal). This is left to a future commit.

These operations should provide a good model for updating the other local
operations.