Add an LZW compression option to GeoTrellis' GeoTIFF encoder.
This feature supports the types supported by the existing GeoTIFF encoder
(Byte, Short, Int, Float, Double). The feature is still experimental and
requires more testing but seems to be working at present.

Future work should focus on extending the types supported by the GeoTIFF
encoder (i.e. 1-bit rasters) as well as support for a "Differencing Predictor"
as outlined in Section 14 of the TIFF 6.0 specification.