Improve color handling/allocation.
This commit touches a lot of pieces of the PNG/color rendering process. The
big idea is to try to avoid the situation where we have a mismatch between the
colors requested by a user and the number of classes to be colored.

There are also some changes which are intended to make book-keeping a little
bit simpler. I had wanted to add some demos/examples of the various ways to
"style" a raster but those will have to wait.