Resolve issues with tiled rasters.
The tiles were being generated with the origin tile placed in the lower
right position (like TMS) but the specification is for the origin tile to
be placed in the upper left.

This commit also adds a new map/reduce style operation class with a
build in threshold for how many operations to run in parallel.