Add in Convolution Operation
Adds an operation that will convolve a given point spread fn
(kernel) with a raster. It also include a simple Gaussian
operation to create a raster with a Gaussian distribution

Also includes tests.

This is the first of a few commits that will hopefully replace
the old kernel density with a better tested implementation