Add a note to Setting up GeoTrellis about catalog.json location
In an attempt to be tidy I tried to store a catalog.json file
for a project in the same directroy as the rasters and metadata files
for that project. This caused a service failure because GeoTrellis
attempted to read the catalog.json as raster metadata.

This commit adds a note to the documentation explaining that the
catalog file must be kept out of the data directory.