Add new simple syntax for creating operations.
This checkin adds a new syntax for transforming a function into an
operation, with the op function(s), and tests.

If you define a function that returns a literal value, an operation to
be executed, or a StepResult (indicating success or failure) the function
can be turned into an operation by passing it to the op function.

For example,

1) A literal value, e.g.

   val plusOne = op { (i:Int) => i + 1 }

2) An operation to be executed:

   val localPlusOne ( (r:Raster, i:Int) => local.Add(r,i + 1) )

3) Or a StepResult (which indicates success or failure)
     val plusOne = op { (i:Int) => Result(i + 1) }