Purge depreciated geometry and FeatureSet code.
In preparation for the 0.8 release, the depreciated geometry API
has been removed along with any code that was dependent on it.

The postgres and postgis related classes are being removed, as we will
now suggest that users use geoslick to load features.

The FeatureSet classes have also been removed as they are not
currently being used in our feature support, but we should
consider bringing them back in the future as a potential
optimization.