Add operation to create histograms from double values.
FastMapHistogram only works with integer values, which is good for performance
but means that, in order to use FastMapHistogram with double values, each value
must be multiplied by a power of ten to preserve significant fractional digits.
This checkin adds an operation and a static method to make this easier.

For example, if you want to save one significant digit (2.1 from 2.123), set
sigificantDigits to 1, and the histogram will save 2.1 as "21".

This does require the user to  sure that the maximum value in the raster multiplied
10 ^ significantDigits does not overflow Int.MaxValue (2,147,483,647).