Replace logic.Reducer1 with logic.TileReducer1.
logic.Reducer1 is an earlier implementation of an abstract base class
for operations that can apply map/reduce style operations on tiled rasters.

It is replaced with TileReducer1, which is similar to ThoroughputLimitedReducer1.
Moving forward, we should remove both Reducer1 and ThoroughputLimitedReducer1,
as well as implement TileReducerN operations, but that isn't included in this
commit.