Make GeoTrellis actor system a singleton.
This commit turns the GeoTrellis akka actor system into a singleton for
purposes of the clustering mechanism.  The actor system is given a single
name ("GeoTrellis") and the tests are changed to no longer create many
servers.  The TestServer class is removed entirely, and the tests are updated
to import from testutils.

In the future, the actor system could be made specific to each server by
having the ServerActor pass its context system into the new server, but that
is not currently implemented.