Update Do operations to reflect operation conventions.
The DoCell family of operations are updated to use the helper raster operations
(for tile unification) as well as to accept their function argument as a second
argument list.  This required eliminating  DoCell apply() method that accepts a
function on double values.

DoDoubleCell was renamed to DoCellDouble, to reflect naming conventions.