Unify focal operations on tiled and untiled rasters.
This commit removes the need to use TileFocalOp when applying focal operations
to tiled rasters.  Focal operations now extend FocalOperation[Raster] with CanTile.
The public analysis area interface has been removed from the base classes.

The current mechanism uses reflection.  I have an initial implementation of an alternate
strategy, but this checkin extends the current strategy to simplify the API for the user.