Refactor tile loading to use LoadRasterTile.
This checkin adds a new operation, LoadRasterTile, to ensure that
the operation generated by TileSetRasterData is not a closure over
the TileSetRasterData itself -- which would cause the data to be
sent over the network when the operation is remotely executed.

LoadUncachedTiledSet is removed because this refactor makes it impossible
distribute work on a tile set that is not loaded from the catalog (unless
the data is sent over the network as a TileArrayRasterData).

This checkin also tests this machinery on the cached tiled raster in the
test collateral.