Extend WebRunner & JettyServer API for custom init parameters.
JettyServer withPackages now accepts an optional initParameters argument for additional
Jetty parameters that may be required by jersey resources specified in the configuration.

The instantiation of WebRunner now includes a call to createJettyServer that can be overwritten
by a subclass of WebRunner to include custom initialization parameters.

In particular, this change is designed to make it possible to use external jetty resources for
service documentation (e.g. swagger's jetty/jaxrs module).