Changed how 'Context' was being used.
Context was simply functionality for operations to load layers. I moved the responsibility of owning a context away from the _run function of operations (since the large majority of ops were not using it). Now if an operation needs to load a layer it can return a LayerResult which takes a function that, given a LayerLoader (what was the Context) will give a final result.

Got rid of junit.

Brought in the GeoTrellis object from the server project to Core.

Renamed get -> run and run -> get. This is how it was in the GeoTrellis object, and I think it makes more sense...from a 'run' you get a 'OperationResult' (was CalculationResult), but if you simply call 'get' on the operation or data source you get the bare result (with an exception thrown if there is an error).