Created CanBuildFrom from ValueSource[_] -> RasterSource
This is for mapping a ValueSource to a raster, for instance if you
rasterize a polygon, you should get a RasterSource back.

Changed the .info on RasterSourceLike to check if the definition is
for a catalogued raster; if so, pull the info from the catalog, else
build an info off of the raster definition.