Create gaussian with parameters in map units.
This commit adds a method to create a gaussian kernel with a
RasterExtent and additional parameters in map units instead of
pixels.  This is useful because, in practice, you almost always
want to define the kernel in terms of map units -- so this reduces
the busy work (as first noted in the ahinz kernel density demo).