Make mkfs.ext3 step idempotent
Before this change set, the mkfs step would blindly attempt to format
the specified devices without checking if they had already been
formatted. Now there is a check that uses `blkid` to determine if
the file system is formatted. The mkfs step now only attempts to format
the device if `blkid` doesn't report ext3 as the device's file system.