Add support for additional projections in geotiffs
This commit improves support for projections when reading and writing
GeoTiffs. Before this commit, it looks as though almost all projections
were read as custom user projections, ignoring and not setting the
`ProjectedCSTypeGeoKey` properly in the process. In the suggestions for
geocoding rasters in the GeoTiff spec, if this code is set (e.g. an EPSG
code) then other parameters are unecessary.

This commit adds support for reading/writing the `ProjectedCSTypeGeoKey`
value properly to enable reading/writing projections like EPSG:3857 (web
-mercator) properly. Tests have been added for this purpose.

In the process of writing this task, I found
http://download.osgeo.org/pub/geotiff/samples/README which describes
additional GeoTiffs GeoTrellis should aim to support reading/writing
with to be on par with GDAL.