Add Option[KeyBounds[K]] Parameter to Layer Writer
A new Option[KeyBounds[K]] parameter has been added to the write method
of the layer writer superclass (Writer.scala).  The motivation here is
to allow the user to write into a key range other than that of the rdd
that is being written (so that key space can be preallocated).