SaveToS3Methods Improvement
Instead of making the S3Client serializable, a companion object of the
method extension class has been created, and that object contains an
apply method that takes a function which returns an S3 Client.  The
extension method supplies a function which returns an actual S3
Client, while the tests supply a mock one.