Tests and Bug Fixes
Tests have been added for various "Double" versions of histogram-related
methods.

Also, bugs in the StreamingHistogram implementation have been fixed: The
DeltaCompare class has been made serializable and registered with Kryo,
a bug in the itemCount method has been corrected so that it gives the
right answer when there are zero or one buckets, and unintentional
sharing of the buckets and deltas when new instances are created has
been eliminated.