Remove constructor type parameters from tile readers
This change makes tile readers consistent with layer readers in that
final types need to be provided to the read method rather than in the
constructor. This also means the tile readers may no longer extent
`Reader[LayerId, Reader[K, V]]` but instead get `TilerReader[LayerId]`
as their super. Again this is consistent with layer readers.