PR #1444 Review: Improve NODATA Handling
Since all bands of Geotrellis MultibandTiles must all have the same
cellType and NODATA value, that particular type is not a good candidate
for representing the data in a GeoTools GridCoverage2D.  The best analog
of a GeoTools GridCoverage2D is an Array of Geotrellis Rasters.

Also, the converter now emits *ConstantNoDataCellType when it can.