(osm2vt) Checkpoint: Lines fuse into Polygons
- Multipolygon handling is an involved process, where some
members of the Relations might be unclosed Ways which only
through that Relation can be realized as a large Polygon.
Collections of unfused Lines can be big (up to 300) and the
Lines can be long (up to 2000 points) so we need to be wary
of performance when attempting to fuse.