Use Cubic Resampling Throughout Tile
Rather than falling back to bilinear interpolation on the edges and
corners, function values are now extrapolated (in nearest-neighbor
fashion) so that derivatives can be computed around the periphery.  Of
course, this will cause the derivatives to decay rapidly the further one
gets from the interior of the tile, but it works better than the
previous behavior.