See https://github.com/jpolchlo/geotrellis/tree/fix/delaunay-stitcher for commmits by jpolchlo
Old PR comments:

This PR aims to fix triangulation problems that arise when attempting to perform distributed triangulations.

These errors affected usage of StitchedDelaunay when applied to BoundaryDelaunay objects. The trouble appears to be related to having an insufficient buffer of triangles in the construction of BoundaryDelaunay instances. It used to be the case that a BoundaryDelaunay object contained all and only the triangles from some source triangulation with circumcircles that extended outside a specified Extent; the "and only" clause has now been relaxed. The criterion is now that the triangle's circumcircle leaves the given extent, or the triangle shares an edge with a triangle that satisfies that condition (or the triangle has an edge which is on the boundary of the triangulation).

This takes Justin's delaunay stitcher fixes, and adds refactors to improve performance of delaunay creation and rasterization.