Relax safeguards on delete methods
Prior to this change, delete methods threw in any case where the
reference to a layer happened to be faulty. This is based on metadata
which, though always present in case of tile data (and necessary for
its deletion), might not be present. Lacking this metadata, we should
log information about the failure and delete any _attribute data we can.