Optimize RasterizeRDD.fromGeometry small intersections
Expect that for large layers a geometry will intersect small fraction of the tiles.
In such a case keeping a set of keys is more efficient than a bitmap.
For small layer it doesn't matter what choice we made.
If this ever fails the next step is a BloomFilter instead of a Set.