Make GeoKeyDirectory parsing more fault tolerant
This commit makes the reading of GeoKeyDirectory entries more fault tolerant to
cover cases where the `NumberOfKeys` entry in the directory header has an
incorrect number of headers that is actually larger than the number of keys.
This manifested itself when attempting to read a tiff created by a geoprocessing
service from an ArcGIS server. The tiff could be read and parsed by other
tools (GDAL, QGIS) without issue but when attempting to open it with GeoTrellis
a `scala.MatchError` would be thrown in `readGeoKeyEntry`.

The workaround introduced here is that if a geokey is read and fails to match
then it likely means there are no longer any valid geokeys to read. I'm open to
other approaches to the workaround than the one offered here - one option is to
introduce a try/catch and only catch on `java.nio.Exception` which can happen if
the number of keys is large enough that reading too many keys (and ignoring
them) leads to a buffer underflow.

Signed-off-by: Chris Brown <cbrown@azavea.com>