Remove SplitMethods[Raster[V]] constraint from spark Split
There is a deeper potential issue here, that `SplitMethods[V]` are not being built up the same way they were before. That should be addressed either in conversion to simulacrum type class encoding or during release testing.
Unclear if its a problem in general, this commit at least cleans up the signature at cost of a little code duplication.