Include feature id in MVT protobuf serialization
Replaces use of geotrellis.vector.Feature with new
geotrellis.vectortile.MVTFeature. MVTFeature aligns
correctly with the expected MVT spec in that it now
includes an Option[Long] identifier as a first class
citizen. Mapping features and geometries to MVTFeature
is left as an application level concern.

Adds tests to verify that feature id and attributes
are serialized to/from protobuf correctly.

Signed-off-by: CloudNiner <andy@finkmiller.com>