Add sparse stitch methods to StitchCollectionMethods
Pulled in from geotrellis/geotrellis-contrib.

The method uses a different algorithm, and therefore
it feels more correct to name it differently from
the original `stitch(): Raster[V]`. Also added a helper
so that users who want to use the collection extent
don't have to know how to get it, since this is the
most likely default use case.