Fix windowed reproject window construction
Investigating strange behavior in windowed reprojection, it was
realized that the gridbounds being used to construct a window was
also being used to construct the expected RasterExtent of said window.
This is not desirable because the unclamped window can be reasonably
expected to be smaller than the grid used to construct it. This entails
more expected cells (and therefore a smaller cell size) than is
desirable. Cross referencing our result with GDAL validated this
concern. See: https://github.com/locationtech/geotrellis/pull/3095