Fix null exceptions in S3Layer classes
...when performing operations against a
bucket with no prefix, e.g.
"s3://example-bucket".

Fixed by preventing AmazonS3URI.getKey from
returning nulls. This _should_ be safe as a
null path prefix is functionally equivalent
to an empty string in that they both represent
"no path prefix".

This change was made to reduce the chance that
nulls escape downstream into our various
Layer[Writer|Reader|Copier|...] methods that
don't have a consistent way to handle a null
prefix.

We also check for null in S3AttributeStore as
a layered defense. More aggressive checking in
each of the downstream S3Layer classes would
be prohibitive ATM because the bucket/prefix
information is copied in each. Users should
prefer use of the apply method over the class
constructor in order to take advantage of this
safety check.