updated computeResolution in IterativeCostDistance to match IterativeViewshed
The compute resolution method in Iterative Cost Distance was calcultating the
cell size based on the length of a degree at the equator without scaling for
the current latitude. This introduces an error as you move away from the
equator. The same method in Iterative Viewshed compensates for this and
maintains greater spatial accuracy.

See issue 3103 https://github.com/locationtech/geotrellis/issues/3103

Signed-off-by: jmtaysom <james.taysom@gmail.com>