FeatureExtractor returns an Iterator instead of Array
Array of features from a sizable can produce very high load on heap, likely blowing out a spark job if it is used in that context. The expectation is that this feature will be used in context of RDD map/flatMap where Iterator is very benficial or the first operation will be a filter or foldLeft on the generated features.