Use GridExtent constructors in combine
So that we're consistently using round
to estimate cols and rows based on cell size.

Remove use of ceilWithTolerance as it's now unclear
if it helps or hinders us in the places its still
used. It can always be re-added in case we hit edge
cases that still fail later.